Fix failing test after Weld 2.4.1.Final upgrade
Weld 2.4.1.Final behaves somewhat different and is initializing the beans
during discovery. That leads to a call
CustomTaskPersistenceContextManager#beginCommandScopedEntityManager
which throws the exception. The bean is thus not initialized and the CDI
container is not able to inject some members (as they were not
initialized and thus are not available for injection). Using the
CommandBasedTaskService mock should lead to a same result - we do assert
that the custom bean is being injected as the execute() methods throws
exception.