[KOGITO-2766] Refactoring initialization code (#763)
Generated handlers are now passed to constructor explicitly (handlers
fields is
removed), which registers them, both for injection and non injection
paths

Configure and registerListeners methods are only generated if they have
content (if there are listeners)

If injection is not used, activate method will be invoked from
constructor. If it is active, then it is invoked using quarkus start
event or springboot post construct as it is now.